+++ ChangeLog.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+## [2.5.3] - 2021-04-23
+
+- improved performances of `VPackParser.toJson()`
 - added `VPackSlice.getSchemaDescription()` to return a printable schema description of the VPack slice
 
 ## [2.5.2] - 2021-03-23